setup/export_embeddings/pyproject.toml
[project]
name = "export-embeddings"
version = "0.1.0"
description = "Export Neo4j chunks with embeddings to JSONL"
requires-python = ">=3.11"
dependencies = [
    "neo4j>=5.0.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
